Added autocommands for allowing wrapping in some files
When I'm writing a commit message or a markdown file it's kind of nice to not have to mess with the auto newline stuff. That can be turned off with `:set wrap` and `:set textwidth=0` but it's tedious to do that manually every time so here's an autocmd group!

I know a lot of people like the auto line break stuff, but I only really look at commit messages on github which have a max width already, and with the auto line break stuff in commits not lining up with the width of the container it's in makes it really distracting. Plus if I rephrase a commit message while I'm writing it I have to redo the line breaks to keep the lines even and I just don't like doing that.