Enable spellcheck on commit messages and markdown files.
I previously just had the wrapping settings enabled for those files, but I wanted spellcheck too. Spellchecking is disabled elsewhere because with treesitter it was complaining about basically everything, so I'm limiting it to files where I'm more likely to make spelling mistakes. I also switched it all to be in a function so I didn't need to duplicate `autocmd` for every line, and I switched to u sing FileType because it seemed more applicable. I also finally figured out what the two file types I needed were which was why I didn't do that in the past.