Setup for Lua
Added lua auto-formatting with stylua, and the sumneko lua lsp. The lua LSP isn't just another random node module to install so I disabled it unless it's actually installed (checks to see if the executable is in the path).

Also rewrote the plugin settings files in lua, as they were generally just all lua code but stuffed in a `.vim` file anyway so it's really not much different than what it was already. The only big change here was rewriting the rest of the autocommands with lua.