Replaced vim-sneak with lightspeed
It seems much more ergonomic. I hadn't been using vim-sneak much because it was a little inconvenient. I would generally need to skip to something that was a repeated word and hitting ; several times to get to the right instance wasn't very convenient. This seems to do a better job.