Added a key map that opens a telescope file finder in the current directory
I often need to find related files in the same directory like an angular template when viewing the matching TS file. This should make it easier to switch between files in the same directory, when it might be otherwise difficult to find them in nvim-tree because the window isn't wide enough to see the file extensions.