Re-added my plugins
Now instead of commenting out some lines loading my plugins with dein `plugins.lua` will now use the local clone of either of my plugins if it exists (on my main computer) or otherwise use the github version anywhere else. This lets me share the same config between devices without having a perpetually dirty file.

Also added packer to the list of plugins, or it'll try and uninstall itself when doing a sync.

Packer seems to generate a file, it's now on the gitignore list.