Transparent terminal and different style for kitty tabs
This changes kitty configuration to use a pink background on the active tab, and a different separator style that I think looks nicer. Also sets the opacity to 0.95 for a slightly transparent terminal.

I had to disable the background on some highlight groups otherwise the terminal wouldn't still be transparent in nvim.

I also had to manually set the neovim title, because with a kitty update the kitty tabs would just show `nvim` instead of the file path I was in. This makes nvim set its window title to the directory name.