Don't include $HOME/code as part of the fzf searching
This changes the fd command to search `$HOME/code` as its base directory instead of telling it to search in there, this returns results that have that part of the path trimmed off so the `$HOME/code` part of the path doesn't get considered part of the search. It gets added back when doing `cd`.