Switch snippet and comment plugins, and keymap changes
Switching to Luasnip and comment.nvim.

Telescope seems to have changed the default for git_files to not include untracked files, so need to pass an argument to get those back or every file needs to be staged before it can be found. While I was at it I rewrote keymap config for Telescope/Harpoon/Projectlaunch.nvim/Markdownmap to lua.