Made some custom commands for copying the buffer or a file at a path
The command `CopyBuf` copies the text of the current buffer to the clipboard.

The command `CopyFile` takes a file path and copies the contents of the file to the clipboard.

The command `CopyFileOnKey` takes a file path and creates a `<leader>C` key binding which will copy that file. Good for if you need to copy the same file multiple times (like copying the output of a command or a generated file).